D_PAD::BuildEffectiveShapes(): fix incorrect rect shape size for rect pads rotated by +-90 deg Noticeable in PnS router.
Fixes #4771
https://gitlab.com/kicad/code/kicad/issues/4771

+++ common/tool/zoom_menu.cpp
@@ -63,7 +63,7 @@ void ZOOM_MENU::update()
 
     const std::vector<double>& zoomList = m_parent->config()->m_Window.zoom_factors;
 
-    for( int i = 0; i < zoomList.size(); ++i )
+    for( size_t i = 0; i < zoomList.size(); ++i )
     {
         // Search for a value near the current zoom setting:
         double rel_error = std::fabs( zoomList[i] - zoom ) / zoom;
